finopti-platform/sub_agents/monitoring_agent_adk/requirements.txt
google-adk>=1.21.0
google-generativeai>=0.8.0
requests>=2.32.4
python-dotenv>=1.0.1
requests>=2.32.4
gunicorn>=21.2.0
python-dotenv>=1.0.0
google-cloud-secret-manager>=2.16.0
google-cloud-bigquery>=3.0.0
deprecated
Flask>=3.0.0
arize-phoenix>=4.0.0
openinference-instrumentation-google-adk>=0.1.0
opentelemetry-sdk
opentelemetry-exporter-otlp
